programação: fix template and repeat directive
Here we go back to using ng-repeat instead of colletion repeat.
RATIONALE: The data we are handling don't need screen fine tunning
(instead, we can choose) the number of records to show and wait for
infinite event to load more data.

The changes needed on the template are already made. CSS needed